+++ client/src/Pages/Forms/Form 10.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import { useState } from 'react';
+import { useState } from "react";
+import {useNavigate} from 'react-router-dom';
 import './Forms.css';
 import { Container } from 'react-bootstrap';
 import Footer from '../../Components/Footer/Footer';
@@ -8,149 +8,47 @@ import NavBar from '../../Components/Navigation Bar/NavBar Student';
 import CancelModal from '../../Components/Modal/Cancel Modal';
 import SubmitModal from '../../Components/Modal/Submit Modal';
 import { fontSize } from '@mui/system';
-
-import { addFormInformation } from './Forms API Call';
-import { uploadImage } from "./Upload Image";
-import { useNavigate } from "react-router-dom";
+import { uploadPdf } from "./Upload Pdf";
+import { addFormInformation } from "./Forms API Call";
 
 
-// Report of Grades
+// Transfer from other schools or UP Units to Cebu
 const Form10 = ({userId}) => {
-    const [price, setPrice] = useState(50);
-    const [image, setImage] = useState()
-    const [formDetails, setFormDetails] = useState({
-        user_id: userId,
-        form_id: 10,
-        remarks: null,
-        student_id: 1,
-        last_name: "",
-        first_name: "",
-        middle_initial: "",
-        student_number: "",
-        mobile_number: "",
-        year_level: "",
-        degree_program: "",
-        email: "",
-        academic_year: "",
-        semester: "",
-        num_copies: "",
-        purpose: "",
-    });
 
     const navigate = useNavigate();
-    
-    const navigateLanding = () => navigate('/student');
-    
-    const handleChange = (e) => {
-        const { name, value } = e.target;
-        console.log(value)
-
-        if (name==="num_copies"){
-            if (value){
-                setPrice(50*parseInt(value))
-            }
-            else{
-                setPrice(50)
-            }
-        }
-        setFormDetails(prevState => ({
-        ...prevState,
-        [name]: value
-        }));
-
-        console.log(formDetails)
-    }
-
+    const classOfferingForm = () => window.location.href = 'https://our.upcebu.edu.ph/wp-content/uploads/2020/05/UPC-FORM-3.1-Application-for-Transfer-from-Other-Schools-or-UP-Units.pdf';
     const [isOpen, setIsOpen] = useState(false);
-    const [isCancelOpen, setIsCancelOpen] = useState(false);
-
-    async function addInfo() {
-        // setIsClicked(true);
-        if (!formValid()) {
-            return;
+    const [pdf, setPdf] = useState()
+    const [formDetails, setFormDetails] = useState({
+            user_id: userId,
+            form_id: 10,
+        });
+    
+    
+        const navigateLanding = () => navigate('/student');     
+    
+        async function addInfo (e) {
+            e.preventDefault(); // Prevent form submission
+            if (pdf) {
+              const formData = new FormData();
+              formData.append('pdf', pdf);
+              formData.append('user_id', formDetails.user_id);
+              const response = addFormInformation(formDetails);
+              uploadPdf(formData);
+              console.log(response);
+              setIsOpen(false);
+              navigateLanding();
+            } else {
+              // Show an alert if no file is uploaded
+              alert('Please upload a file.');
+            }
+          }
+    
+        const pdfHandler = (e) => {
+            const file = e.target.files[0];
+            console.log(file)
+            setPdf(file)
         }
-
-        const formData = new FormData()
-        formData.append('image', image)
-        formData.append('user_id', formDetails.user_id)
-        const response = addFormInformation(formDetails);
-        uploadImage(formData)
-        console.log(response)
-        setIsOpen(false)
-        navigateLanding()
-
-    }
-
-    const formValid = () => {
-        const {
-            last_name,
-            first_name,
-            student_number,
-            mobile_number,
-            year_level,
-            degree_program,
-            email,
-            academic_year,
-            semester,
-            num_copies,
-            purpose
-        } = formDetails;
-
-        if (
-            !last_name ||
-            !first_name ||
-            !student_number ||
-            !mobile_number ||
-            !year_level ||
-            !degree_program ||
-            !email ||
-            !academic_year ||
-            !semester ||
-            !num_copies ||
-            !purpose
-        ) {
-            // Form validation failed
-            alert("Please fill in all fields");
-            return false;
-        }
-
-        if (isNaN(student_number) || isNaN(mobile_number)) {
-            alert("Student number and mobile number must be integers.");
-            return;
-        }
-
-        if (!image) {
-            // No file selected for upload
-            alert("Please select a file to upload");
-            return false;
-        }
-
-        return true;
-    };
-
-    const pdfHandler = (e) => {
-        const file = e.target.files[0];
-        console.log(file)
-        setImage(file)
-    }
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        setIsOpen(true);
-    };
-
-    const handleCancel = (e) => {
-        e.preventDefault();
-        setIsCancelOpen(true);
-    };
-
-    const handleCancelModalClose = () => {
-        setIsCancelOpen(false);
-    };
-
-    const handleSubmitModalClose = () => {
-        setIsOpen(false);
-    };
 
     return(
         <div>
@@ -161,117 +59,30 @@ const Form10 = ({userId}) => {
             </div>
             <Container>
                 <div className="form-title">
-                    Report of Grades
+                    Transfer from other schools or UP Units to Cebu
                 </div>
-                <form class="tcg-form"  onSubmit={handleSubmit}>
-                    <h1 className='form-group-title'>A. Student Details</h1>
-                    <div class="form-row">    
-                        <div class="col-md-3 mb-2">     
-                            <label for="studentLastName">Last Name</label>
-                            <input type="text" class="form-control" id="studentLastName" name="last_name" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-3 mb-2">     
-                            <label for="studentFirstName">First Name</label>
-                            <input type="text" class="form-control" id="studentFirstName" name="first_name" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-2 mb-2">     
-                            <label for="studentMiddleInitial">Middle Initial</label>
-                            <input type="text" class="form-control" id="studentMiddleInitial" name="middle_initial" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-2 mb-2">
-                            <label for="studentNumber">Student Number</label>
-                            <input type="text" class="form-control" id="studentNumber" name= "student_number" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-2 mb-2">
-                            <label for="mobileNumber">Mobile Number</label>
-                            <input type="text" class="form-control" id="mobileNumber" name="mobile_number" onChange={(e) => handleChange(e)}/>
-                        </div>
-                    </div>
-                    <div class="form-row">
-                        <div class="col-md-6 mb-2">
-                            <label for="degreeProgram">Degree Program</label>
-                            <select class="custom-select" id='degreeProgram' name="degree_program" onChange={(e) => handleChange(e)}>
-                            <option value="BS Computer Science">BS Computer Science</option>
-                                <option value="BS Biology">BS Biology</option>
-                                <option value="BS Mathematics">BS Mathematics</option>
-                                <option value="BS Statistics">BS Statistics</option>
-                            </select>
-                        </div>
-                        <div class="col-md-2 mb-2">
-                            <label for="yearLevel">Year Level</label>
-                            <input type="number" min='1' max='6' class="form-control" id="yearLevel" name="year_level" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-4 mb-2">
-                            <label for="emailAddress">Email Address</label>
-                            <input type="email" class="form-control" id="emailAddress" name="email_address" onChange={(e) => handleChange(e)}/>
-                        </div>
-                    </div>
-                    <h1 className='form-group-title'>B. Request Details</h1>
-                    <div class="form-row">
-                        <div class="col-md-4 mb-2">     
-                            <label for="academicYear">Academic Year</label>
-                            <input type="text" class="form-control" id="academicYear" name="academic_year" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-4 mb-2">
-                            <label for="semester">Semester</label>
-                            <input type="text" class="form-control" id="semester" name="semester" onChange={(e) => handleChange(e)}/>
-                        </div>
-                        <div class="col-md-4 mb-2">
-                            <label for="copies">Number of Copies</label>
-                            <input type="text" class="form-control" id="copies" name="num_copies" onChange={(e) => handleChange(e)}/>
-                        </div>
-                    </div>
-                    <div class="form-row">
-                        <div class="col-md-7 mb-2">
-                            <label for="purpose">Purpose for Request</label>
-                            <select class="custom-select" id='purpose' onChange={(e) => handleChange(e)}>
-                                <option selected value=""> </option>
-                                <option value="scholarship">Scholarship</option>
-                                <option value="readmission">Readmission</option>
-                                <option value="shift">Shift</option>
-                            </select>
-                        </div>
-                    </div>
-                    <h1 className='form-group-title'>C. Proof of Payment</h1>
-                    <h2 className='form-subtitle-1'>Pay via Philippine Veterans Bank</h2>
-                    <div className="inline-form-details">
-                        <div className='form-details'><b>Account Name:</b> UP Cebu</div>
-                        <div className='form-details'><b>Account Number:</b> 21-0000-067</div>
-                        <div className='form-details'><b>Branch:</b> Cebu City</div>
-                    </div>
-                    <h3 className='form-subtitle-2'>Note:</h3>
-                    <div className='form-details'>If asked for mobile number, please provide your own mobile number</div>
-                    <div className='form-details'>If paying via GCash, please add 4 zeroes at the start to have 13 digits, 000021-0000-067.</div>
-                    <h3 className='form-subtitle-2'>Total Amount to be Paid</h3>
-                    <div className="request-price-container">
-                        <div className="column-1">
-                            <div className='form-details-price'>₱{price}.00</div>
-                            <div className='form-details'>This is the total amount to be paid through Philippine Veterans Bank</div>
-                            <div className='form-details'>(Payment may be made via online channels such as gcash, instapay, pesonet, bank transfers.)</div>
-                        </div>
-                        <div className="column-2">
-                            <div class="form-group">
-                                <input type="file" class="form-control-file" id="paymentProof" name="image" accept="image/*" multiple={false} onChange={pdfHandler}/>
-                            </div>
+                <form class="tcg-form"  onSubmit={addInfo}>
+                    <h1 className='form-group-title'>A. Request Details</h1>
+                    <div className="form-description-text">
+                            <p className='form-description-text -1'>Applicants shall have completed, outside the University, at least thirty-three (33) academic units and must have a general weighted average of at least 2.0 or better for all the collegiate academic units taken outside of UP.</p>
+                            <p className='form-description-text -2'>Kindly download and fill up this form: </p>                     
+                     </div>
+                     <div>
+                        <p className='download-form' onClick={classOfferingForm} >APPLICATION FORM FOR TRANSFER FROM OTHER SCHOOL(S)/OTHER UP UNIT(S)  </p>                     
+                     </div>
+
+                     <div className="form-description-text">
+                            <p className='form-description-text-3'>Upload the filled up request form here: </p>                     
+                     </div>
+
+
+                    <div className="upload">
+                        <div class="form-group">
+                            <input type="file" class="form-control-file" id="paymentProof" name="pdf" accept="application/pdf" multiple={false} onChange={pdfHandler}/>
                         </div>
                     </div>
 
-                    <div className="form-row">
-                    <div class="col-md-4 mb-2">
-                            <label for="paymentMode">Mode of Payment</label>
-                            <select class="custom-select" id="paymentMode">
-                                <option selected value="1">GCASH</option>
-                                <option value="2">Bank</option>
-                            </select>
-                        </div>
-                    </div>
-                    <h3 className='form-subtitle-2'>Important:</h3>
-                    <ul className='important-list'>
-                        <li className='important-item'>Payment status must be COMPLETED or SUCCESSFUL.</li>
-                        <li className='important-item'>Proof of payment (screenshot/image) must include <b><u>Amount Paid</u></b>, <b><u>Transaction Date</u></b>, <b><u>Time</u></b>, and <b><u>Reference Number</u></b>.</li>
-                        <li className='important-item'>For GCash transactions, make sure to enable mobile or email notifications. Attach proof of completed payment from GCash SMS or GCash Email with InstaPay Trace Number. (also found in your GCash Inbox, "Your Bank Transfer Status" message)</li>
-                        <li className='important-item'>Otherwise, if payment is not verifiable, your request may not be processed or may be delayed.</li>
-                    </ul>
+
                     <div className='privacy-notice-container'>
                         <h1 className="form-group-title">Privacy Notice for UP Students</h1>
                         <div className="privacy-notice-text">
@@ -285,13 +96,13 @@ const Form10 = ({userId}) => {
                     <div className="form-buttons-container">
                     <div className="cancel-button">
                         <button class="btn btn-primary" type="submit" onClick={() => setIsOpen(true)}>Cancel</button>
-                        {isCancelOpen && <CancelModal setIsOpen={setIsCancelOpen} />}
+                        {isOpen && <CancelModal setIsOpen={setIsOpen} />}
                     </div>
                     <div className="submit-button">
                         <button class="btn btn-primary" onClick={() => setIsOpen(true)}>Submit</button> 
-                        {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo}/>}
-                    </div>
-                    </div>
+                        {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo} />}
+                    </div> 
+                </div> 
             </Container>
             <Footer/>
         </div>
