+++ client/src/Pages/Forms/Form 16.jsx
@@ -12,43 +12,45 @@ import { uploadPdf } from "./Upload Pdf";
 import { addFormInformation } from "./Forms API Call";
 
 
-// Justification for Non-dissolution of Small Class Size
+// Change in Class Offering
 const Form16 = ({userId}) => {
 
     const navigate = useNavigate();
     const classOfferingForm = () => window.location.href = 'https://our.upcebu.edu.ph/wp-content/uploads/2022/02/UPC-FORM-Request-for-Change-in-Class-Offerings-Fillable.pdf';
     const [isOpen, setIsOpen] = useState(false);
-    const [pdf, setPdf] = useState()
+  
+    const [pdf, setPdf] = useState(null);
     const [formDetails, setFormDetails] = useState({
-            user_id: userId,
-            form_id: 16,
-        });
-    
-    
-        const navigateLanding = () => navigate('/student');     
-    
-        async function addInfo (e) {
-            e.preventDefault(); // Prevent form submission
-            if (pdf) {
-              const formData = new FormData();
-              formData.append('pdf', pdf);
-              formData.append('user_id', formDetails.user_id);
-              const response = addFormInformation(formDetails);
-              uploadPdf(formData);
-              console.log(response);
-              setIsOpen(false);
-              navigateLanding();
-            } else {
-              // Show an alert if no file is uploaded
-              alert('Please upload a file.');
-            }
-        }
-    
-        const pdfHandler = (e) => {
-            const file = e.target.files[0];
-            console.log(file)
-            setPdf(file)
-        }
+      user_id: userId,
+      form_id: 16,
+    });
+  
+    const navigateLanding = () => navigate('/student');
+  
+    async function addInfo (e) {
+      e.preventDefault(); // Prevent form submission
+      if (pdf) {
+        const formData = new FormData();
+        formData.append('pdf', pdf);
+        formData.append('user_id', formDetails.user_id);
+        const response = addFormInformation(formDetails);
+        uploadPdf(formData);
+        console.log(response);
+        setIsOpen(false);
+        navigateLanding();
+      } else {
+        // Show an alert if no file is uploaded
+        alert('Please upload a file.');
+      }
+    }
+  
+    const pdfHandler = (e) => {
+      const file = e.target.files[0];
+      console.log(file);
+      setPdf(file);
+    };
+  
+  
 
     return(
         <div>
@@ -59,19 +61,22 @@ const Form16 = ({userId}) => {
             </div>
             <Container>
                 <div className="form-title">
-                    Justification for Non-dissolution of Small Class Size
+                    Change in Class Offering
                 </div>
                 <form class="tcg-form" onSubmit={addInfo}>
                     <h1 className='form-group-title'>A. Request Details</h1>
-                    <ol>
-                        <li>The student makes a letter of appeal for late registration/ payment addressed to the University Registrar or the Chancellor. </li>
-                        <li>The student emails his/her letter to the Program Adviser and Division Chair. The Program Adviser and the Division Chair affix their e-signature on the student’s letter.</li>
-                        <li>Upload the approved letter</li>
-                    </ol>
+                    <div className="form-description-text">
+                            <p className='form-description-text -1'>Once pre-enlistment has started, any changes made to the class offerings for the specified term, would require a written request with approval.</p>
+                            <p className='form-description-text -2'>Kindly download and fill up this form: </p>                     
+                     </div>
+                     <div>
+                        <p className='download-form' onClick={classOfferingForm} >REQUEST FOR CHANGE IN CLASS OFFERINGS  </p>                     
+                     </div>
+
 
                     <div className="upload">
                         <div class="form-group">
-                            <input type="file" class="form-control-file" id="paymentProof" name="pdf" accept="application/pdf" multiple={false} onChange={pdfHandler}/>
+                            <input type="file" class="form-control-file" id="paymentProof" name="pdf" accept="pdf" multiple={false} onChange={pdfHandler}/>
                         </div>
                     </div>
 
@@ -85,8 +90,8 @@ const Form16 = ({userId}) => {
                             <p className='privacy-notice-text-end'>"I hereby certify that all information given above are true and correct."</p>
                         </div>
                     </div>
-                    </form>
-                    <div className="form-buttons-container">
+                </form>
+                <div className="form-buttons-container">
                     <div className="cancel-button">
                         <button class="btn btn-primary" type="submit" onClick={() => setIsOpen(true)}>Cancel</button>
                         {isOpen && <CancelModal setIsOpen={setIsOpen} />}
