+++ client/src/Pages/Forms/Form 2.jsx
@@ -15,14 +15,43 @@ import { useEffect } from "react";
 
 // Certification of GWA
 const Form2 = ({userId}) => {
+        const [image, setImage] = useState();
         const [price, setPrice] = useState(50);
-        const [image, setImage] = useState()
+
+        const [savedDetails, setSavedDetails] = useState({})
+
+        useEffect(() => {
+            async function getDetails(data){
+                const response = await axios.get('http://localhost:5000/student_api/getDetails/'+ data)
+                console.log(response.data[0])
+                setSavedDetails(response.data[0])
+            }
+    
+            getDetails(userId)
+        },[])
+
+        useEffect(() => {
+            console.log(formDetails)
+            setFormDetails({
+                user_id: userId,
+                form_id: 2,
+                last_name: (savedDetails.last_name) ? savedDetails.last_name : "",
+                first_name: (savedDetails.first_name) ? savedDetails.first_name : "",
+                middle_initial: (savedDetails.middle_initial) ? savedDetails.middle_initial : "",
+                student_number: (savedDetails.student_number) ? savedDetails.student_number : "",
+                mobile_number: (savedDetails.mobile_number) ? savedDetails.mobile_number : "",
+                year_level: (savedDetails.year_level) ? savedDetails.year_level : "",
+                degree_program: (savedDetails.degree_program) ? savedDetails.degree_program : "",
+                email: (savedDetails.email) ? savedDetails.email : "",
+            })
+            console.log(formDetails)
+        },[savedDetails])
+    
         const [formDetails, setFormDetails] = useState({
-            user_id: userId,
+            user_id: userId.toString(),
             form_id: 2,
             remarks: null,
-            student_id: 1,
-            last_name: "",
+             last_name: "",
             first_name: "",
             middle_initial: "",
             student_number: "",
@@ -35,112 +64,112 @@ const Form2 = ({userId}) => {
             num_copies: "",
             purpose: "",
         });
-        
-        const [savedDetails, setSavedDetails] = useState({})
 
-        async function getDetails(data){
-            const response = await axios.get('http://localhost:5000/student_api/getDetails/'+ data)
-            console.log(response.data[0])
-            setSavedDetails(response.data[0])
-        }
-
-        useEffect(() => {
-            getDetails(userId)
-        },[])
 
         const navigate = useNavigate();
-    
+
         const navigateLanding = () => navigate('/student');
-        
+    
         const handleChange = (e) => {
             const { name, value } = e.target;
-            console.log(value)
+            setFormDetails(prevState => ({
+                ...prevState,
+                [name]: value
+            }))
+            
+            console.log(formDetails);
     
-            if (name==="num_copies"){
-                if (value){
-                    setPrice(50*parseInt(value))
-                }
-                else{
-                    setPrice(50)
+            if (name === "num_copies") {
+                if (value) {
+                    setPrice(50 * parseInt(value));
+                } else {
+                    setPrice(50);
                 }
             }
-            setFormDetails(prevState => ({
-            ...prevState,
-            [name]: value
-            }));
+        };
     
-            console.log(formDetails)
-        }
     
         const [isOpen, setIsOpen] = useState(false);
         const [isCancelOpen, setIsCancelOpen] = useState(false);
     
         async function addInfo() {
+
+
             if (!formValid()) {
                 return;
             }
     
             const formData = new FormData();
             formData.append('image', image);
-            formData.append('user_id', formDetails.user_id);
-            const response = addFormInformation(formDetails);
+            formData.append('user_id', userId);
+            console.log (formDetails);
+            const response = await addFormInformation(formDetails);
+            formData.append('id', response.data)
+            console.log(response.data)
             uploadImage(formData);
             setIsOpen(false);
             navigateLanding();
         }
+
+        const formValid = () => {
+            const {
+              last_name,
+              first_name,
+              student_number,
+              mobile_number,
+              year_level,
+              degree_program,
+              email,
+              academic_year,
+              semester,
+              num_copies,
+              purpose
+            } = formDetails;
+          
+            const emptyFields = [];
+          
+            if (!last_name) emptyFields.push("Last Name");
+            if (!first_name) emptyFields.push("First Name");
+            if (!student_number) emptyFields.push("Student Number");
+            if (!mobile_number) emptyFields.push("Mobile Number");
+            if (!year_level) emptyFields.push("Year Level");
+            if (!degree_program) emptyFields.push("Degree Program");
+            if (!email) emptyFields.push("Email");
+            if (!academic_year) emptyFields.push("Academic Year");
+            if (!semester) emptyFields.push("Semester");
+            if (!num_copies) emptyFields.push("Number of Copies");
+            if (!purpose) emptyFields.push("Purpose");
+          
+            if (emptyFields.length > 0) {
+              // Form validation failed
+              const errorMessage = `Please fill in the following fields:\n${emptyFields.join(
+                ", "
+              )}`;
+              alert(errorMessage);
+              return false;
+            }
+          
+            if (isNaN(student_number) || isNaN(mobile_number)) {
+              alert("Student Number and mobile number must be integers.");
+              return;
+            }
+          
+            if (!image) {
+              // No file selected for upload
+              alert("Please select a file to upload");
+              return false;
+            }
+          
+            return true;
+          };
+          
+
     
         const pdfHandler = (e) => {
             const file = e.target.files[0];
             console.log(file)
             setImage(file)
         }
-
-        const formValid = () => {
-            const {
-                last_name,
-                first_name,
-                student_number,
-                mobile_number,
-                year_level,
-                degree_program,
-                email,
-                academic_year,
-                semester,
-                num_copies,
-                purpose
-            } = formDetails;
-    
-            if (
-                !last_name ||
-                !first_name ||
-                !student_number ||
-                !mobile_number ||
-                !year_level ||
-                !degree_program ||
-                !email ||
-                !academic_year ||
-                !semester ||
-                !num_copies ||
-                !purpose
-            ) {
-                // Form validation failed
-                alert("Please fill in all fields");
-                return false;
-            }
-    
-            if (isNaN(student_number) || isNaN(mobile_number)) {
-                alert("Student number and mobile number must be integers.");
-                return;
-            }
-    
-            if (!image) {
-                // No file selected for upload
-                alert("Please select a file to upload");
-                return false;
-            }
-    
-            return true;
-        };
 
         const handleSubmit = (e) => {
             e.preventDefault();
@@ -159,7 +188,8 @@ const Form2 = ({userId}) => {
         const handleSubmitModalClose = () => {
             setIsOpen(false);
         };
-    
+        {console.log(savedDetails.degree_program)}
+
     
 
     return(
@@ -178,30 +208,30 @@ const Form2 = ({userId}) => {
                     <div class="form-row">    
                         <div class="col-md-3 mb-2">     
                             <label for="studentLastName">Last Name</label>
-                            <input type="text" class="form-control" id="studentLastName" name="last_name" onChange={(e) => handleChange(e)} defaultValue={savedDetails.last_name}/>
+                            <input type="text" class="form-control" id="studentLastName" name="last_name" onChange={(e) => handleChange(e)} defaultValue={savedDetails.last_name} key={savedDetails.last_name}/>
                         </div>
                         <div class="col-md-3 mb-2">     
                             <label for="studentFirstName">First Name</label>
-                            <input type="text" class="form-control" id="studentFirstName" name="first_name" onChange={(e) => handleChange(e)} defaultValue={savedDetails.first_name}/>
+                            <input type="text" class="form-control" id="studentFirstName" name="first_name" onChange={(e) => handleChange(e)} defaultValue={savedDetails.first_name} key={savedDetails.first_name}/>
                         </div>
                         <div class="col-md-2 mb-2">     
                             <label for="studentMiddleInitial">Middle Initial</label>
-                            <input type="text" class="form-control" id="studentMiddleInitial" name="middle_initial" onChange={(e) => handleChange(e)} defaultValue={savedDetails.middle_initial}/>
+                            <input type="text" class="form-control" id="studentMiddleInitial" name="middle_initial" onChange={(e) => handleChange(e)} defaultValue={savedDetails.middle_initial} key={savedDetails.middle_initial}/>
                         </div>
                         <div class="col-md-2 mb-2">
                             <label for="studentNumber">Student Number</label>
-                            <input type="text" class="form-control" id="studentNumber" name= "student_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.student_number}/>
+                            <input type="text" class="form-control" id="studentNumber" name= "student_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.student_number} key={savedDetails.student_number}/>
                         </div>
                         <div class="col-md-2 mb-2">
                             <label for="mobileNumber">Mobile Number</label>
-                            <input type="text" class="form-control" id="mobileNumber" name="mobile_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.mobile_number}/>
+                            <input type="text" class="form-control" id="mobileNumber" name="mobile_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.mobile_number} key={savedDetails.mobile_number}/>
                         </div>
                     </div>
                     <div class="form-row">
                     <div class="col-md-6 mb-2">
                         <label for="degreeProgram">Degree Program</label>
-                            <select class="custom-select" id='degreeProgram' name="degree_program" onChange={(e) => handleChange(e)}>
-                                <option selected value=""> </option>
+                            <select class="custom-select" id='degreeProgram' name="degree_program" defaultValue={savedDetails.degree_program}  key={savedDetails.degree_program} onChange={(e) => handleChange(e)}>
+                            <option selected defaultValue={savedDetails.degree_program}> </option>
                                 <option value="BS Computer Science">BS Computer Science</option>
                                 <option value="BS Biology">BS Biology</option>
                                 <option value="BS Mathematics">BS Mathematics</option>
@@ -210,11 +240,11 @@ const Form2 = ({userId}) => {
                         </div>
                         <div class="col-md-2 mb-2">
                             <label for="yearLevel">Year Level</label>
-                            <input type="number" min='1' max='6' class="form-control" id="yearLevel" name="year_level" onChange={(e) => handleChange(e)}  defaultValue={savedDetails.year_level}/>
+                            <input type="number" min='1' max='6' class="form-control" id="yearLevel" name="year_level" onChange={(e) => handleChange(e)}  defaultValue={savedDetails.year_level} key={savedDetails.year_level}/>
                         </div>
                         <div class="col-md-4 mb-2">
                             <label for="emailAddress">Email Address</label>
-                            <input type="email" class="form-control" id="emailAddress" name="email_address" onChange={(e) => handleChange(e)} defaultValue={savedDetails.email}/>
+                            <input type="email" class="form-control" id="emailAddress" name="email_address" onChange={(e) => handleChange(e)} defaultValue={savedDetails.email} key={savedDetails.email}/>
                         </div>
                     </div>
                     <h1 className='form-group-title'>B. Request Details</h1>
@@ -235,7 +265,7 @@ const Form2 = ({userId}) => {
                     <div class="form-row">
                         <div class="col-md-7 mb-2">
                             <label for="purpose">Purpose for Request</label>
-                            <select class="custom-select" id='purpose' onChange={(e) => handleChange(e)}>
+                            <select class="custom-select" id='purpose' name="purpose" onChange={(e) => handleChange(e)}>
                                 <option selected value=""> </option>
                                 <option value="scholarship">Scholarship</option>
                                 <option value="readmission">Readmission</option>
@@ -306,8 +336,8 @@ const Form2 = ({userId}) => {
             </Container>
             <Footer/>
         </div>
-    )
-}
+    );
+};
