+++ client/src/Pages/Forms/Form 21.jsx
@@ -6,13 +6,16 @@ import Header from '../../Components/Header/Header';
 import NavBar from '../../Components/Navigation Bar/NavBar Student';
 import CancelModal from '../../Components/Modal/Cancel Modal';
 import SubmitModal from '../../Components/Modal/Submit Modal';
+import { fontSize } from '@mui/system';
+
 import { addFormInformation } from './Forms API Call';
 import { uploadImage } from "./Upload Image";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 
-//True Copy of Grades (TCG)
-const Form1 = ({userId}) => {
+
+// Certification of Underload
+const Form21 = ({userId}) => {
     const [image, setImage] = useState();
     const [price, setPrice] = useState(50);
 
@@ -35,7 +38,7 @@ const Form1 = ({userId}) => {
         console.log(formDetails)
         setFormDetails({
             user_id: userId,
-            form_id: 1,
+            form_id: 21,
             last_name: (savedDetails.last_name) ? savedDetails.last_name : "",
             first_name: (savedDetails.first_name) ? savedDetails.first_name : "",
             middle_initial: (savedDetails.middle_initial) ? savedDetails.middle_initial : "",
@@ -50,7 +53,7 @@ const Form1 = ({userId}) => {
 
     const [formDetails, setFormDetails] = useState({
         user_id: userId.toString(),
-        form_id: 1,
+        form_id: 21,
         remarks: null,
          last_name: "",
         first_name: "",
@@ -181,9 +184,8 @@ const Form1 = ({userId}) => {
     };
     {console.log(savedDetails.degree_program)}
 
-    return (
-        <div>
 
+    return(
         <div>
             <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous"/> 
             <NavBar/>
@@ -192,11 +194,11 @@ const Form1 = ({userId}) => {
             </div>
             <Container>
                 <div className="form-title">
-                    True Copy of Grades (TCG)
+                    Certification of Underload
                 </div>
-                <form class="tcg-form" onSubmit={handleSubmit}>
+                <form class="tcg-form" >
                     <h1 className='form-group-title'>A. Student Details</h1>
-                    <div class="form-row">
+                    <div class="form-row">    
                         <div class="col-md-3 mb-2">     
                             <label for="studentLastName">Last Name</label>
                             <input type="text" class="form-control" id="studentLastName" name="last_name" onChange={(e) => handleChange(e)} defaultValue={savedDetails.last_name} key={savedDetails.last_name}/>
@@ -211,7 +213,7 @@ const Form1 = ({userId}) => {
                         </div>
                         <div class="col-md-2 mb-2">
                             <label for="studentNumber">Student Number</label>
-                            <input type="text" class="form-control" id="studentNumber" name="student_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.student_number} key={savedDetails.student_number}/>
+                            <input type="text" class="form-control" id="studentNumber" name= "student_number" onChange={(e) => handleChange(e)} defaultValue={savedDetails.student_number} key={savedDetails.student_number}/>
                         </div>
                         <div class="col-md-2 mb-2">
                             <label for="mobileNumber">Mobile Number</label>
@@ -235,7 +237,7 @@ const Form1 = ({userId}) => {
                         </div>
                         <div class="col-md-4 mb-2">
                             <label for="emailAddress">Email Address</label>
-                            <input type="email" class="form-control" id="emailAddress" name="email" onChange={(e) => handleChange(e)} defaultValue={savedDetails.email}  key={savedDetails.email}/>
+                            <input type="email" class="form-control" id="emailAddress" name="email_address" onChange={(e) => handleChange(e)} defaultValue={savedDetails.email} key={savedDetails.email}/>
                         </div>
                     </div>
                     <h1 className='form-group-title'>B. Request Details</h1>
@@ -255,8 +257,8 @@ const Form1 = ({userId}) => {
                     </div>
                     <div class="form-row">
                         <div class="col-md-7 mb-2">
-                            <label for="purpose">Purpose for TCG</label>
-                            <select class="custom-select" id='purpose' name="purpose" onChange={(e) => handleChange(e)}>
+                            <label for="purpose">Purpose for Request</label>
+                            <select class="custom-select" id='purpose' onChange={(e) => handleChange(e)}>
                                 <option selected value=""> </option>
                                 <option value="scholarship">Scholarship</option>
                                 <option value="readmission">Readmission</option>
@@ -313,8 +315,8 @@ const Form1 = ({userId}) => {
                             <p className='privacy-notice-text-end'>"I hereby certify that all information given above are true and correct."</p>
                         </div>
                     </div>
-                </form>
-                <div className="form-buttons-container">
+                    </form>
+                    <div className="form-buttons-container">
                     <div className="cancel-button">
                         <button class="btn btn-primary" type="submit" onClick={() => setIsOpen(true)}>Cancel</button>
                         {isCancelOpen && <CancelModal setIsOpen={setIsCancelOpen} />}
@@ -323,14 +325,16 @@ const Form1 = ({userId}) => {
                         <button class="btn btn-primary" onClick={() => setIsOpen(true)}>Submit</button> 
                         {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo}/>}
                     </div>
-                </div>
+                    </div>
             </Container>
-
             <Footer/>
-            </div>
-
         </div>
-    );
-};
+    )
+}
 
-export default Form1;
+
+
+export default Form21;
+
+
+
