+++ client/src/Pages/Forms/Form 6.jsx
@@ -1,5 +1,4 @@
 import { useState } from "react";
-import {useNavigate} from 'react-router-dom';
 import './Forms.css';
 import { Container } from 'react-bootstrap';
 import Footer from '../../Components/Footer/Footer';
@@ -7,49 +6,122 @@ import Header from '../../Components/Header/Header';
 import NavBar from '../../Components/Navigation Bar/NavBar Student';
 import CancelModal from '../../Components/Modal/Cancel Modal';
 import SubmitModal from '../../Components/Modal/Submit Modal';
-import { fontSize } from '@mui/system';
-import { uploadPdf } from "./Upload Pdf";
-import { addFormInformation } from "./Forms API Call";
+//import { fontSize } from '@mui/system';
+import { addFormInformation } from './Forms API Call';
+import { useNavigate } from "react-router-dom";
 
 
-// Late Payment /Extension of Enrolment
+
+// Dropping of Course
 const Form6 = ({userId}) => {
-
-    const navigate = useNavigate();
-    const classOfferingForm = () => window.location.href = 'https://our.upcebu.edu.ph/wp-content/uploads/2022/02/UPC-FORM-Request-for-Change-in-Class-Offerings-Fillable.pdf';
-    const [isOpen, setIsOpen] = useState(false);
-
-    const [pdf, setPdf] = useState()
     const [formDetails, setFormDetails] = useState({
-            user_id: userId,
-            form_id: 6,
-        });
+        user_id: userId,
+        form_id: 6,
+        remarks: null,
+        student_id: 1,
+        last_name: "",
+        first_name: "",
+        middle_initial: "",
+        student_number: "",
+        degree_program: "",
+        year_level: "",
+        subject_dropped: "",
+        section: "",
+        instructor_name: "",
+        purpose: ""
+    });
     
+    const navigate = useNavigate();
+
+    const navigateLanding = () => navigate('/student');
     
-        const navigateLanding = () => navigate('/student');     
-    
-        async function addInfo (e) {
-            e.preventDefault(); // Prevent form submission
-            if (pdf) {
-              const formData = new FormData();
-              formData.append('pdf', pdf);
-              formData.append('user_id', formDetails.user_id);
-              const response = addFormInformation(formDetails);
-              uploadPdf(formData);
-              console.log(response);
-              setIsOpen(false);
-              navigateLanding();
-            } else {
-              // Show an alert if no file is uploaded
-              alert('Please upload a file.');
-            }
-          }
-    
-        const pdfHandler = (e) => {
-            const file = e.target.files[0];
-            console.log(file)
-            setPdf(file)
+    const handleChange = (e) => {
+        const { name, value } = e.target;
+        console.log(value)
+
+        setFormDetails(prevState => ({
+        ...prevState,
+        [name]: value
+        }));
+
+        console.log(formDetails)
+    }
+
+    const [isOpen, setIsOpen] = useState(false);
+    const [isCancelOpen, setIsCancelOpen] = useState(false);
+
+    async function addInfo() {
+        if (!formValid()) {
+            return;
         }
+
+        const formData = new FormData();
+        formData.append('user_id', formDetails.user_id);
+        const response = addFormInformation(formDetails);
+        setIsOpen(false);
+        navigateLanding();
+    }
+
+
+    const formValid = () => {
+        const {
+
+            last_name,
+            first_name,
+            student_number,
+            year_level,
+            degree_program,
+            subject_dropped,
+            section,
+            instructor_name,
+            purpose
+
+        } = formDetails;
+
+        if (
+            !last_name ||
+            !first_name ||
+            !student_number ||
+            !year_level ||
+            !degree_program ||
+            !subject_dropped||
+            !section||
+            !instructor_name||
+            !purpose
+        ) {
+            // Form validation failed
+            alert("Please fill in all fields");
+            return false;
+        }
+
+        if (isNaN(student_number)) {
+            alert("Student number and mobile number must be integers.");
+            return;
+        }
+
+        return true;
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        setIsOpen(true);
+    };
+
+    const handleCancel = (e) => {
+        e.preventDefault();
+        setIsCancelOpen(true);
+    };
+
+    const handleCancelModalClose = () => {
+        setIsCancelOpen(false);
+    };
+
+    const handleSubmitModalClose = () => {
+        setIsOpen(false);
+    };
+
+
+
 
     return(
         <div>
@@ -60,23 +132,65 @@ const Form6 = ({userId}) => {
             </div>
             <Container>
                 <div className="form-title">
-                    Late Payment /Extension of Enrolment
+                    Dropping of Course
                 </div>
-                <form class="tcg-form" onSubmit={addInfo}>
-                    <h1 className='form-group-title'>A. Request Details</h1>
-                    <div className="form-description-text">
-                            <p className='form-description-text -1'>1. The student makes a letter of appeal for late registration/ payment addressed to the University Registrar or the Chancellor. </p>
-                            <p className='form-description-text -2'>2. The student emails his/her letter to the Program Adviser and Division Chair. The Program Adviser and the Division Chair affix their e-signature on the student’s letter. </p>       
-                            <p className='form-description-text -3'>3. Upload the approved letter </p>                     
-                     </div>
-
-
-                    <div className="upload">
-                        <div class="form-group">
-                            <input type="file" class="form-control-file" id="paymentProof" name="pdf" accept="application/pdf" multiple={false} onChange={pdfHandler}/>
+                <form class="tcg-form" onSubmit={handleSubmit}>
+                    <h1 className='form-group-title'>A. Student Details</h1>
+                    <div class="form-row">
+                        <div class="col-md-3 mb-2">     
+                            <label for="studentLastName">Last Name</label>
+                            <input type="text" class="form-control" id="studentLastName" name="last_name"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                        <div class="col-md-3 mb-2">     
+                            <label for="studentFirstName">First Name</label>
+                            <input type="text" class="form-control" id="studentFirstName" name="first_name"  onChange={(e) => handleChange(e)} />
+                        </div>
+                        <div class="col-md-2 mb-2">     
+                            <label for="studentMiddleInitial">Middle Initial</label>
+                            <input type="text" class="form-control" id="studentMiddleInitial" name="middle_initial"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                        <div class="col-md-4 mb-2">
+                            <label for="studentNumber">Student Number</label>
+                            <input type="text" class="form-control" id="studentNumber" name="student_number" onChange={(e) => handleChange(e)}/>
                         </div>
                     </div>
-
+                    <div class="form-row">
+                    <div class="col-md-6 mb-2">
+                        <label for="degreeProgram">Degree Program</label>
+                            <select class="custom-select" id='degreeProgram' name="degree_program" onChange={(e) => handleChange(e)}>
+                                <option selected value=""> </option>
+                                <option value="BS Computer Science">BS Computer Science</option>
+                                <option value="BS Biology">BS Biology</option>
+                                <option value="BS Mathematics">BS Mathematics</option>
+                                <option value="BS Statistics">BS Statistics</option>
+                            </select>
+                        </div>
+                        <div class="col-md-2 mb-2">
+                            <label for="yearLevel">Year Level</label>
+                            <input type="number" min='1' max='6' class="form-control" id="yearLevel" name="year_level"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                    </div>
+                    <h1 className='form-group-title'>B. Request Details</h1>
+                    <div class="form-row">
+                        <div class="col-md-4 mb-2">     
+                            <label for="academicYear">Subject Dropped</label>
+                            <input type="text" class="form-control" id="academicYear" name="subject_dropped"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                        <div class="col-md-4 mb-2">
+                            <label for="semester">Section</label>
+                            <input type="text" class="form-control" id="section" name="section"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                        <div class="col-md-4 mb-2">
+                            <label for="copies">Instructor's Name</label>
+                            <input type="text" class="form-control" id="instructorName" name="instructor_name"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                    </div>
+                    <div class="form-row">
+                    <div class="col-md-12 mb-2">
+                            <label for="copies">Reason</label>
+                            <input type="text" class="form-control" id="purpose" name="purpose"  onChange={(e) => handleChange(e)}/>
+                        </div>
+                    </div>
 
                     <div className='privacy-notice-container'>
                         <h1 className="form-group-title">Privacy Notice for UP Students</h1>
@@ -87,18 +201,17 @@ const Form6 = ({userId}) => {
                             <p className='privacy-notice-text-end'>"I hereby certify that all information given above are true and correct."</p>
                         </div>
                     </div>
-                    </form>
-                    <div className="form-buttons-container">
+                </form>
+                <div className="form-buttons-container">
                     <div className="cancel-button">
                         <button class="btn btn-primary" type="submit" onClick={() => setIsOpen(true)}>Cancel</button>
-                        {isOpen && <CancelModal setIsOpen={setIsOpen} />}
+                        {isCancelOpen && <CancelModal setIsOpen={setIsCancelOpen} />}
                     </div>
                     <div className="submit-button">
                         <button class="btn btn-primary" onClick={() => setIsOpen(true)}>Submit</button> 
-                        {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo} />}
-                    </div> 
-                </div>  
-                
+                        {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo}/>}
+                    </div>
+                </div>
             </Container>
             <Footer/>
         </div>
