+++ client/src/Pages/Forms/Form 8.jsx
@@ -12,16 +12,16 @@ import { uploadPdf } from "./Upload Pdf";
 import { addFormInformation } from "./Forms API Call";
 
 
-// LOA
+// Request for Substitution of Courses/Subjects
 const Form8 = ({userId}) => {
 
     const navigate = useNavigate();
-    const classOfferingForm = () => window.location.href = 'https://our.upcebu.edu.ph/wp-content/uploads/2023/01/UPC-FORM-Application-for-Leave-of-Absence-20230127.pdf';
+    const classOfferingForm = () => window.location.href = 'https://our.upcebu.edu.ph/wp-content/uploads/2017/02/UPC-Substitution-20170526.pdf';
     const [isOpen, setIsOpen] = useState(false);
     const [pdf, setPdf] = useState()
     const [formDetails, setFormDetails] = useState({
             user_id: userId,
-            form_id: 8,
+            form_id: 18,
         });
     
     
@@ -50,7 +50,6 @@ const Form8 = ({userId}) => {
             setPdf(file)
         }
 
-
     return(
         <div>
             <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous"/> 
@@ -60,20 +59,18 @@ const Form8 = ({userId}) => {
             </div>
             <Container>
                 <div className="form-title">
-                    Request for Leave of Absence 
+                    Request for Substitution of Courses/Subjects
                 </div>
-                <form class="tcg-form" onSubmit={addInfo}>
-                    <h1 className='form-group-title'>A. Student Details</h1>
+                <form class="tcg-form" >
+                    <h1 className='form-group-title'>A. Request Details</h1>
                     <div className="form-description-text">
-                            <p className='form-description-text -1'>A student who does not intend to enroll in a semester may apply for a leave of absence (LOA). This also applies to a student currently enrolled and who intends to withdraw his/her enrollment for the rest of the term. A student on leave of absence does not sever their ties with the University. A request for a leave of absence should be made in writing to the Dean. The request should state the reason for the leave and should specify the period. The leave may be approved for a period of one (1) year but may be renewed for at most another year.</p>
-                            <p className='form-description-text -2'>Kindly download and fill up this form: </p>                     
+                            <p className='form-description-text -1'>Kindly download and fill up this form: </p>   
+                            <p className='form-description-text -2'>Secure and fill out the Substitution Form, in three copies.</p>
+                            <p className='form-description-text -3'>Get endorsement from the Instructor of the Subject Required, your Program Adviser, the Department Chair/Program Coordinator of the Subject Required and the Department Chair/Program Coordinator of the Subject Taken.</p> 
+                            <p className='form-description-text -4'>Submit the signed Substitution Form to the Office of the College Secretary.</p>                                         
                      </div>
                      <div>
-                        <p className='download-form' onClick={classOfferingForm} >APPLICATION FOR LEAVE OF ABSENCE (LOA)</p>                     
-                     </div>
-
-                     <div className="form-description-text">
-                            <p className='form-description-text-3'>Upload the filled up request form here: </p>                     
+                        <p className='download-form' onClick={classOfferingForm} >REQUEST FOR SUBSTITUTION OF COURSES/SUBJECT</p>                     
                      </div>
 
 
@@ -93,17 +90,17 @@ const Form8 = ({userId}) => {
                             <p className='privacy-notice-text-end'>"I hereby certify that all information given above are true and correct."</p>
                         </div>
                     </div>
-                    </form>
-                    <div className="form-buttons-container">
+                </form>
+                <div className="form-buttons-container">
                     <div className="cancel-button">
                         <button class="btn btn-primary" type="submit" onClick={() => setIsOpen(true)}>Cancel</button>
                         {isOpen && <CancelModal setIsOpen={setIsOpen} />}
                     </div>
                     <div className="submit-button">
                         <button class="btn btn-primary" onClick={() => setIsOpen(true)}>Submit</button> 
-                        {isOpen && <SubmitModal setIsOpen={setIsOpen} action={addInfo} />}
-                    </div> 
-                </div> 
+                        {isOpen && <SubmitModal setIsOpen={isOpen} />}
+                    </div>
+                </div>
             </Container>
             <Footer/>
         </div>
