+++ server/routes/form.js
@@ -100,8 +100,10 @@ router.post("/transaction_made_upload",async (req,res,err)=> {        //API endp
   if(now.getMinutes() < 10){minute = "0" + now.getMinutes().toString()} else{minute = now.getMinutes().toString()}
   if(now.getSeconds() < 10){second = "0" + now.getSeconds().toString()} else{second = now.getSeconds().toString()}
 
+  console.log(req.body)
   let transaction_id = now.getYear().toString() + months + dates + hour + minute + second + req.body.user_id.toString()
   const formId = req.body.form_id
+
 
   const form_values = await new Promise((resolve) => {
     db.query("SELECT form_duration, form_name FROM forms WHERE form_id = ?", formId, (err, data) => {
@@ -150,6 +152,7 @@ router.post('/transaction_made', async (req,res) =>{        //API endpoint for s
   if(now.getHours() < 10){hour = "0" + now.getHours().toString()} else{hour = now.getHours().toString()}
   if(now.getMinutes() < 10){minute = "0" + now.getMinutes().toString()} else{minute = now.getMinutes().toString()}
 
+  console.log(req.body)
   let transaction_id = now.getYear().toString() + months + dates + hour + minute + req.body.user_id.toString()
   if(req.body.form_id >= 0 && req.body.form_id <= 3){
     q2 = 'INSERT INTO transaction_info (`transaction_id`,`last_name`, `first_name`, `middle_initial`, `student_number`, `mobile_number`, `year_level`, `degree_program`, `email`, `academic_year`, `semester`, `num_copies`, `purpose`) VALUES (?)'
